fix(alchemy): make provider dependencies optional peers
Move provider-specific packages out of the core dependency set so projects install the provider packages matching the alchemy subpaths they import. Lazy-load AWS and Cloudflare CLI provider modules so unrelated CLI commands do not require those optional peer packages at startup.
