+++ examples/aws-ec2/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@distilled.cloud/aws": "catalog:",
     "@effect/platform-node": "catalog:",
+    "@types/aws-lambda": "catalog:",
     "alchemy": "workspace:*",
     "effect": "catalog:"
   }
