+++ examples/aws-static-site/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@distilled.cloud/aws": "catalog:",
     "@effect/platform-node": "catalog:",
+    "@types/aws-lambda": "catalog:",
     "alchemy": "workspace:*",
     "effect": "catalog:"
   }
