+++ examples/cloudflare-solidjs-ssr/package.json
@@ -12,13 +12,20 @@
   },
   "license": "MIT",
   "devDependencies": {
+    "@cloudflare/workers-types": "catalog:",
+    "@distilled.cloud/cloudflare": "catalog:",
+    "@distilled.cloud/cloudflare-rolldown-plugin": "catalog:",
+    "@distilled.cloud/cloudflare-runtime": "catalog:",
+    "@distilled.cloud/cloudflare-vite-plugin": "catalog:",
     "@tailwindcss/postcss": "^4.1.13",
-    "postcss": "^8.4.49",
-    "solid-devtools": "^0.34.3",
-    "tailwindcss": "^4.1.13",
-    "vite": "^7.1.4",
     "alchemy": "workspace:*",
     "effect": "catalog:",
+    "ignore": "^7.0.5",
+    "postcss": "^8.4.49",
+    "solid-devtools": "^0.34.3",
+    "sonda": "catalog:",
+    "tailwindcss": "^4.1.13",
+    "vite": "^7.1.4",
     "vite-plugin-solid": "^2.11.8"
   },
   "dependencies": {
