+++ examples/cloudflare-solidstart/package.json
@@ -10,14 +10,21 @@
     "destroy": "alchemy destroy"
   },
   "dependencies": {
+    "@cloudflare/workers-types": "catalog:",
+    "@distilled.cloud/cloudflare": "catalog:",
+    "@distilled.cloud/cloudflare-rolldown-plugin": "catalog:",
+    "@distilled.cloud/cloudflare-runtime": "catalog:",
+    "@distilled.cloud/cloudflare-vite-plugin": "catalog:",
     "@solidjs/meta": "^0.29.4",
     "@solidjs/router": "^0.15.0",
     "@solidjs/start": "2.0.0-alpha.2",
-    "solid-js": "^1.9.5",
+    "@solidjs/vite-plugin-nitro-2": "^0.1.0",
     "alchemy": "workspace:*",
     "effect": "catalog:",
-    "vite": "^7.0.0",
-    "@solidjs/vite-plugin-nitro-2": "^0.1.0"
+    "ignore": "^7.0.5",
+    "solid-js": "^1.9.5",
+    "sonda": "catalog:",
+    "vite": "^7.0.0"
   },
   "engines": {
     "node": ">=22"
