+++ examples/cloudflare-worker-async/package.json
@@ -19,10 +19,17 @@
   "dependencies": {
     "@alchemy.run/better-auth": "workspace:*",
     "@cloudflare/workers-types": "catalog:",
+    "@distilled.cloud/cloudflare": "catalog:",
+    "@distilled.cloud/cloudflare-rolldown-plugin": "catalog:",
+    "@distilled.cloud/cloudflare-runtime": "catalog:",
+    "@distilled.cloud/cloudflare-vite-plugin": "catalog:",
     "@effect/platform-bun": "catalog:",
     "@effect/platform-node": "catalog:",
     "alchemy": "workspace:*",
     "better-auth": "catalog:",
-    "effect": "catalog:"
+    "effect": "catalog:",
+    "ignore": "^7.0.5",
+    "sonda": "catalog:",
+    "vite": "catalog:"
   }
 }
